refactor: migrate aws_route_53_resolver_endpoint to new structure
